embulk-jruby-strptime: Separate RubyDateParser from StrptimeParser implementation
